+++ bin/make_examples_const.dart
@@ -68,7 +68,7 @@ void rewrite(
     newSource.write(originalSource);
   } else {
     var modifiedSource = originalSource.replaceFirst(
-        examples.group(1), examples.group(1) + 'const');
+        examples.group(1), '${examples.group(1)}const');
     newSource.write(modifiedSource);
   }
 }