+++ test/generate_localized/code_map_messages_fr.dart
@@ -3,27 +3,32 @@
 // messages from the main program should be duplicated here with the same
 // function name.
 
-import 'package:intl/message_lookup_by_library.dart';
+// Ignore issues from commonly used lints in this file.
+// ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
+// ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
+// ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
+// ignore_for_file:unused_import, file_names
 
+import 'package:intl/intl.dart';
+import 'package:intl/message_lookup_by_library.dart';
+import 'dart:convert';
 import 'code_map_messages_all.dart' show evaluateJsonTemplate;
 
-final messages = MessageLookup();
+import 'dart:collection';
 
-typedef MessageIfAbsent = String Function(String messageStr, List<Object> args);
+final messages = new MessageLookup();
+
+typedef String MessageIfAbsent(String messageStr, List<Object> args);
 
 class MessageLookup extends MessageLookupByLibrary {
-  @override
   String get localeName => 'fr';
 
-  @override
-  // ignore: type_annotate_public_apis
-  String evaluateMessage(translation, List<dynamic> args) {
+  String evaluateMessage(Object translation, List<dynamic> args) {
     return evaluateJsonTemplate(translation, args);
   }
 
-  @override
   Map<String, dynamic> get messages => _constMessages;
   static const _constMessages = <String, Object>{
-    'Hello from application': "Bonjour de l'application"
+    "Hello from application": "Bonjour de l'application"
   };
 }