+++ test/message_extraction/embedded_plural_text_before_test.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+@Timeout(Duration(seconds: 180))
+
 library embedded_plural_text_before_test;
 
 import 'package:test/test.dart';