+++ test/message_extraction/failed_extraction_test.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+@Timeout(Duration(seconds: 180))
+
 library failed_extraction_test;
 
 import 'dart:io';
@@ -11,8 +13,8 @@ import 'package:test/test.dart';
 import 'message_extraction_test.dart';
 
 void main() {
-  test('Expect warnings but successful extraction', () {
-    runTestWithWarnings(warningsAreErrors: false, expectedExitCode: 0);
+  test('Expect warnings but successful extraction', () async {
+    await runTestWithWarnings(warningsAreErrors: false, expectedExitCode: 0);
   });
 }
 
@@ -21,12 +23,11 @@ const List<String> defaultFiles = [
   'part_of_sample_with_messages.dart'
 ];
 
-void runTestWithWarnings({
-  bool warningsAreErrors,
-  int expectedExitCode,
-  bool embeddedPlurals = true,
-  List<String> sourceFiles = defaultFiles,
-}) {
+Future<void> runTestWithWarnings(
+    {bool warningsAreErrors,
+    int expectedExitCode,
+    bool embeddedPlurals = true,
+    List<String> sourceFiles = defaultFiles}) async {
   void verify(ProcessResult result) {
     try {
       expect(result.exitCode, expectedExitCode);
@@ -35,8 +36,9 @@ void runTestWithWarnings({
     }
   }
 
-  copyFilesToTempDirectory();
-  var program = 'bin/extract_to_arb.dart';
+  await copyFilesToTempDirectory();
+
+  var program = asTestDirPath('../../bin/extract_to_arb.dart');
   List<String> args = ['--output-dir=$tempDir'];
   if (warningsAreErrors) {
     args.add('--warnings-are-errors');
@@ -44,13 +46,11 @@ void runTestWithWarnings({
   if (!embeddedPlurals) {
     args.add('--no-embedded-plurals');
   }
-  List<String> allArgs = [
-    ...args,
-    ...sourceFiles,
-  ];
+  var files = sourceFiles.map(asTempDirPath).toList();
+  List<String> allArgs = [program, ...args, ...files];
   var callback = expectAsync1(verify);
 
-  run(program, allArgs).then(callback);
+  run(null, allArgs).then(callback);
 }
 
 typedef ThenArgument = dynamic Function(ProcessResult _);