+++ test/message_extraction/mock_flutter/foo_messages_de_DE.dart
@@ -6,6 +6,6 @@ library keep_the_static_analysis_from_complaining;
 
 class MessageLookup {
   String get messages => throw UnimplementedError(
-      'This entire file is only here to make the static'
-      ' analysis happy. It will be generated during actual tests.');
+    'This entire file is only here to make the static'
+    ' analysis happy. It will be generated during actual tests.');
 }