+++ test/message_extraction/run_and_verify.dart
@@ -6,9 +6,8 @@ library verify_and_run;
 import 'dart:convert';
 import 'dart:io';
 
-// TODO(devoncarew): See the comment below about restoring this testing.
-// import 'sample_with_messages.dart' as sample;
-// import 'verify_messages.dart';
+import 'sample_with_messages.dart' as sample;
+import 'verify_messages.dart';
 
 void main(List<String> args) async {
   if (args.isEmpty) {
@@ -16,21 +15,18 @@ void main(List<String> args) async {
     exit(0);
   }
 
-  // TODO(devoncarew): This disables the verification of the generated code.
-  // We'll want to restore testing it, but likely move to a more hygenic
-  // integration style test, or, use a technique like golden masters.
-  // // Verify message translation output
-  // await sample.main();
-  // verifyResult();
+  // Verify message translation output
+  await sample.main();
+  verifyResult();
 
-  // Messages with skipExtraction set should not be extracted.
+  // Messages with skipExtraction set should not be extracted
   var fileArgs = args.where((x) => x.contains('.arb'));
-  Map<String, dynamic> messages =
-      jsonDecode(File(fileArgs.first).readAsStringSync());
+  var messages = jsonDecode(File(fileArgs.first).readAsStringSync())
+      as Map<String, dynamic>;
   messages.forEach((name, _) {
     // Assume any name with 'skip' in it should not have been extracted.
     if (name.contains('skip')) {
-      throw 'A skipped message was extracted: $name';
+      throw "A skipped message was extracted ('$name')";
     }
   });
 }